(actually) finish rename
